feat: add 3 modul
Переделал проект, добавил 3-й (basic) модуль и 2 зависимых (core, api).
Проект запускается без сборки (она не обязательна).
