--- api/src/main/java/ru/clevertec/App.java
@@ -1,12 +1,0 @@
-package ru.clevertec;
-
-import ru.clevertec.util.Utils;
-
-public class App {
-    public static void main(String[] args) {
-
-        System.out.println("Utils.isAllPositiveNumbers( args ) = "
-                + Utils.isAllPositiveNumbers("12", "79"));
-
-    }
-}
