+++ basic/api/build.gradle
@@ -10,6 +10,12 @@ repositories {
 }
 
 dependencies {
+
+    implementation 'org.slf4j:slf4j-api:1.7.32'
+    implementation 'ch.qos.logback:logback-classic:1.4.11'
+
+    implementation project(":basic:core")
+
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
 }
