+++ basic/api/src/main/java/ru/clevertec/App.java
@@ -1,0 +1,12 @@
+package ru.clevertec;
+
+import ru.clevertec.utils.Util;
+
+public class App {
+    public static void main(String[] args) {
+
+        System.out.println("Utils.isAllPositiveNumbers( args ) = "
+                + Util.isAllPositiveNumbers("12", "79"));
+
+    }
+}
