+++ basic/build.gradle
@@ -1,19 +1,29 @@
 plugins {
     id 'java'
+    id 'jacoco'
 }
 
 group = 'ru.clevertec'
-version = '1.0-SNAPSHOT'
+version = '1.3.5'
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
+
+    implementation 'org.slf4j:slf4j-api:1.7.32'
+    implementation 'ch.qos.logback:logback-classic:1.4.11'
+
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
 }
 
+apply from: 'scriptPlugin.gradle'
+
 test {
     useJUnitPlatform()
 }
+
+test.finalizedBy jacocoTestReport
+jacocoTestReport.dependsOn test
