--- core/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'ru.clevertec'
-version = '1.3.5'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-
-    implementation files('libs/utils-1.3.5.jar')
-
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
-test {
-    useJUnitPlatform()
-}
