--- core/src/main/java/ru/clevertec/util/Utils.java
@@ -1,17 +1,0 @@
-package ru.clevertec.util;
-
-import ru.clevertec.stringUtil.StringUtils;
-
-public class Utils {
-
-    public static boolean isAllPositiveNumbers(String... str) {
-
-        for (String s : str) {
-            if (!StringUtils.isPositiveNumber(s)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-}
