+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Fri Oct 13 12:54:59 MSK 2023
+#Wed Oct 18 13:16:54 MSK 2023
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-8.0-bin.zip
